Add a bunch of derived variables.
The new variables are:
- divergence
- vorticity
- eddy kinetic energy
- lapse rate
- total column vapor
- total column liquid
- total column ice
- integrate vapor transport
- relative humidity

These variables may be useful for evaluating the physical consistency of forecast models.

I have also taken the opportunity to slightly adjust the interface for DerivedVariable, by removing the required variable_name field. I don't see a use-case for keeping track of this variable names outside of dictionary keys.

PiperOrigin-RevId: 565820289