Add (a)geostrophic wind speed derived variable
As suggested by https://arxiv.org/abs/2309.08473

Example geostrophic wind on ERA5, which is well defined everywhere except near the equator:
![image](https://github.com/google-research/weatherbench2/assets/1217238/c4d7d6f3-c8ef-4ed1-a646-f570027ef066)

Example ageostrophic wind, which is is an order of magnitude smaller than the geostrophic wind in the extra-tropics:
![image](https://github.com/google-research/weatherbench2/assets/1217238/fc6df3c9-ae58-43b4-b57b-5a231aa38587)

Geostrophic wind only makes sense to evaluate in the extra-tropics, because it is undefined/infinite on the equator. Ensuring that it can be evaluated in the extra-tropics also required a small tweak to our spatial averaging code.

PiperOrigin-RevId: 570265149