Adjust windows so that `method=resample` and `method=rolling` align. Also set `skipna=False` when computing statistics.  This is anticipation of the following development workflow: * Make '1w' rolling window version of ground truth   * Compute '1w' rolling window version of 0.25degree ERA5, 1990-2023, starting from 1h spaced ERA   * Regrid and compute derived variables   * Compute climatology * Make resampled forecasts   * Forecast to 6-16 week lead times, with 12hr timedelta spacing, and init time spacing that is an odd multiple of 12h.   * resample forecasts to weekly and delete (ttl) original forecasts to save space * Evaluate resampled forecasts vs. rolled ground truth
Using `rolling` for ground truth is better than `resample`. If we used `resample`, then we'd have to be sure our forecast start times and timedeltas were aligned with the window used for ground truth

Clearly we could also roll the forecasts, but that barely increases statistical power and wastes time & space. For final WB evals, that may make sense, but for development we want something faster and more lightweight.

PiperOrigin-RevId: 631443536